+++ src/mew/string.cpp
@@ -66,7 +66,7 @@ namespace
 		{
 			if(m_capacity >= capacity)
 				return;
-			m_capacity = math::max(16U, capacity, m_capacity*2);
+			m_capacity = math::max(static_cast<size_t>(16), capacity, m_capacity*2);
 			const size_t bytesize = sizeof(Struct) + m_capacity * sizeof(T);
 			m_data = (T*)((UINT8*)::realloc(data(), bytesize) + sizeof(Struct));
 		}
